adding models/schema; not actually a CI model
If anyone has a better guide of where to put something like this, please
let me know. As is, though, it's the basic SQL schema for the tables
which run the site. They're not listed anywhere else, and should be
packaged with the rest of the source code.

You will want to create and be using the database (i.e., CREATE SCHEMA
dm; USE dm;) before running the statements, else it'll create tables in
some random place.