big change; removing jQuery for most pages
this was partly an exercise, but by the end jQuery is only here for the
flot library. Otherwise, the JS has all been rewritten to use the native
HTML5 functions. Older Internet Explorer, etc. will not have as great of
a time, here, but I feel like everything is standardized enough at this
point that it's time to go in that direction.