Introducing TERRAIN
This commit was a very long time coming. Many months ago, I started work
on a rebuild of this application, creating a framework of my own, from
scratch, requiring PHP 5.4 and above.

I finished the core platform in a few weeks, though I've let it sit
before releasing it to the world. I've used this same framework, which
I'm currently calling TERRAIN, in a number of other projects, and it's
held together rather perfectly.

PHP isn't quite the hip kid on the block that it was years ago, but as a
language it has legs, and is extremely efficient and productive. It's
shocking how few lines of code get a fully working system up and
running. Everything you need is here. The core language itself goes a
long way, and the point of this framework is to showcase that fact.

This application is also using Backbone.js, which with a few minor
thought-adjustments works quite well natively with PHP.

Finally, I'm releasing this and TERRAIN under the MPL v2.0. That's
important for me, as TERRAIN is designed to be built-to-fit. I'd rather
have a library of implementations that we all can share than a system
made cumbersome by its attempts at covering infinitely many cases. If
you don't like how one part works, change just that and release it. Keep
your own development private, even; just make sure that any changes to
core files are shared with the world.