fix never create new `user` re: lack of type
each element in the $columns array needs to have value set for the
'type' index, otherwise we pass it by and go onto the next one, in the
set method. Here, as we had no other columns to set, we had nothing to
do, and the new user was never created.